+++ packages/create-codes/templates/react/src/app.tsx
